requirements.txt
APScheduler==3.6.3
async-timeout==4.0.3
cachetools==4.2.2
certifi==2023.7.22
charset-normalizer==3.4.1
idna==3.10
importlib-metadata==6.7.0
python-telegram-bot==13.7
pytz==2025.1
redis==5.0.8
requests==2.31.0
six==1.17.0
tornado==6.2
typing_extensions==4.7.1
tzdata==2025.1
tzlocal==5.1
urllib3==1.26.18
wincertstore==0.2
zipp==3.15.0
standard-imghdr==3.13.0